requirements.txt
python-vlc
requests